Edit dr embed if in last msgs

+++ cogs/deathroll.py
@@ -141,9 +141,15 @@ class DeathrollButton(discord.ui.Button['DeathRoll']):
                 embed = view.get_deathroll_end_embed()
             view.stop()
 
-        await interaction.response.defer()
-        await interaction.message.delete()
-        await interaction.channel.send(content=None, embed=embed, view=view)
+        # Check if Deathroll message is in last X messages of the channel
+        channel = interaction.message.channel
+        history = [message async for message in channel.history(limit=4)]
+        if interaction.message in history:
+            await interaction.response.edit_message(content=None, embed=embed, view=view)
+        else:
+            await interaction.response.defer()
+            await interaction.message.delete()
+            await interaction.channel.send(content=None, embed=embed, view=view)
 
 
 class DeathRoll(discord.ui.View):
